requirements.txt
torch
torchvision
pillow
streamlit
pyngrok
pydicom
opencv-python
matplotlib
grad-cam
ultralytics
